connpool: make call() faster for specific cases
The connection pool `call()` function used to connect to all of the
instances matching some specified static requiriments and synchroniouly
select one of the active. It was slower a lot times slower than its
possible because the connpool only actually needs one active connection
to perform the call.

This patch makes the connection pool `call()` function work a bit
smarter. Now if the priorities (`prefer_ro`/`prefer_rw` call modes)
aren't used within the `connpool.call()` the pool will wait for any
connection matching the requirements as a candidate to perform the call.

In case of using the priorities all of the candidates matching static
properties are fetched. It may be optimized using the fact that
technically it's possible to determine whether the reached replica has
the maiximum priority and can be used without waiting other ones.

Closes #10598

NO_DOC=performance improvement, no API change