config: support bs=native for failover=off
This is very similar to the `replication.failover` = `manual` mode, but
takes into account that it is possible that several RW instances appear
simultaneously.

On a replicaset bootstrap the first RW instance (lexicographically)
takes the bootstrap leader role (using the graceful bootstrap
procedure).

During the replicaset lifetime (after the bootstrap) the bootstrap
leadership is taken when an instance goes to RW. Any RW instance may
take this role, not only the first RW one. The last one that goes to RW
wins.

The problematic scenario is when a current bootstrap leader goes off or
becomes inaccessible. Another RW instance doesn't pick up this role
automatically and some RW -> RO -> RW configuration changes are needed
to make another instance take the role. In this regard, the strategy
seems worse than `auto`.

Part of #10857

NO_DOC=to be added in the latest patch of the series
NO_CHANGELOG=same