config: allow guest to perform failover.execute
This patch grants the guest user runtime privileges for running
`failover.execute` function. This is an internal function used by the
supervised failover coordinator. This privileges are needed for the
guest user to be able to perform the initial bootstrap since there is no
other users available before it.

Technically it's a part of a native and supervised bootstrap strategy
support allowing the external agent (i.e the supervised failover
coordinator) to perform the initial bootstrap.

Needed for tarantool/tarantool-ee#1194

NO_CHANGELOG=internal change
NO_TEST=will be tested in EE
NO_DOC=internal change